socle_transverse_4.0.17/socle-transverse-ui/src/main/java/fr/dila/st/ui/bean/actions/STLockActionDTO.java
package fr.dila.st.ui.bean.actions;

public class STLockActionDTO {
    private boolean currentDocIsLockActionnableByCurrentUser;

    public STLockActionDTO() {}

    public boolean getCurrentDocIsLockActionnableByCurrentUser() {
        return currentDocIsLockActionnableByCurrentUser;
    }

    public void setCurrentDocIsLockActionnableByCurrentUser(boolean currentDocIsLockActionnableByCurrentUser) {
        this.currentDocIsLockActionnableByCurrentUser = currentDocIsLockActionnableByCurrentUser;
    }
}
